Get an element from tensor

+++ tests/test_tensor_shape.cpp
@@ -2,7 +2,7 @@
 #include <cassert>
 
 int main() {
-  std::vector<int> shape = {2, 2};
+  std::vector<size_t> shape = {2, 2};
   transformer::Tensor<int> tensor(shape);
   assert(tensor.shape() == shape);
 }
